sql int 오류 수정

+++ src/main/kotlin/com/example/demo/content/tag/AttachedTagKey.kt
@@ -7,8 +7,8 @@ import java.io.Serializable
 @Embeddable
 data class AttachedTagKey(
     @Column(name = "content_id", nullable = false)
-    val contentId: Long = 0,
+    val contentId: Int = 0,
 
     @Column(name = "tag_id", nullable = false)
-    val tagId: Long = 0
+    val tagId: Int = 0
 ) : Serializable
